Started work on moving BLE communication to Dashboard Page
Because the BLE connection work is an ongoing process, we want to tie it to an activity. It's also too much work to get the BLE info back to the main activity, so I'm setting up a dashboard activity. Once I get the dashboard working with BLE, I'll implement the application flow back in. At this point, every time you re-resume the app, you have to re-connect to the device.