Cleaned up setting up stream/connecting to device. IT LIGHTS UP :joy:
for real, I cleaned up the logic for starting streaming, and since I noticed that the service wasn't connecting to the device, I cleaned up that logic as well. The device does start generating data, but it also disconnects as soon as you connect to it. I need to fix the logic to maintain the connection, but I'm just happy that it does connect.