Some debugging work on trying to get notifications.
I believe that the issue right now is that while we are setting a notification on the device, we still have to write a value to the descriptor of the characteristic that we want to check. If you run my code as is, you will find that I cannot initiate the write command to write to that descriptor. This is where I am stuck, because I cannot figure out why.