Worked on UI for adding a second device
Most of these changes involved the changing of BtButtonActivity so that when you press the button, the items are still displayed on the screen as a list of checkable things. Now, you can select one or two devices to connect to. *AS OF THIS CURRENT COMMIT*, the system ignores the second device being connected to. That will be involved in the next part of all of this.