Finished BLE Backend- Receives characteristics fff2-fff5
Through a long and arduous battle with semaphores, I have ensured that for each of the two devices that the application is connecting to the device, looking at each characteristic, writing the descriptor, and starting the stream by writing the value. This has been tested and confirmed to work- to the best of my knowledge. I am getting all 4 packets, but the UI is not updating, mostly due to some error in the HR calculation.

There are a lot of added lines in the service because I had to assign a unique callback object for each device. Apparently, the system does not generate a new callback object when the device sends out data, rather, the callback object is populated with data and alerts the app. Each of the two callbacks is programmed to deal with the specific device it is designed for. If anyone wants, and I don't think you do, you can try to make a factory to make callback objects and can deal with generic input. I made it work with generic input, in theory.